Python Full Analytic/Python requirement.txt
# requirements.txt
# Core dependencies
python>=3.11
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
torch>=2.0.0
transformers>=4.35.0

# Data & Database
asyncpg>=0.29.0
sqlalchemy>=2.0.0
redis>=5.0.0
pymongo>=4.5.0

# API & Communication
grpcio>=1.59.0
grpcio-tools>=1.59.0
aiohttp>=3.9.0
websockets>=12.0
requests>=2.31.0

# Data Processing
polars>=0.19.0
pyarrow>=14.0.0
dask>=2023.10.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Monitoring
prometheus-client>=0.19.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.4.0
click>=8.1.0
rich>=13.5.0
tenacity>=8.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0